refactor(orc8r): Segregation of Tenants servicer as internal
