+++ orc8r/cloud/go/services/tenants/obsidian/handlers/handlers.go
@@ -23,8 +23,8 @@ import (
 	"magma/orc8r/cloud/go/obsidian"
 	"magma/orc8r/cloud/go/services/tenants"
 	"magma/orc8r/cloud/go/services/tenants/obsidian/models"
+	"magma/orc8r/cloud/go/services/tenants/protos"
 	"magma/orc8r/lib/go/errors"
-	"magma/orc8r/lib/go/protos"
 )
 
 const (
