+++ orc8r/cloud/go/services/tenants/protos/tenants.proto
@@ -1,0 +1,63 @@
+/*
+Copyright 2020 The Magma Authors.
+
+This source code is licensed under the BSD-style license found in the
+LICENSE file in the root directory of this source tree.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+syntax = "proto3";
+
+import "orc8r/protos/common.proto";
+
+package magma.orc8r;
+option go_package = "magma/orc8r/cloud/go/services/tenants/protos";
+
+message GetTenantRequest {
+    int64 id = 1;
+}
+
+message Tenant {
+    string name = 1;
+    repeated string networks = 2;
+}
+
+message TenantList {
+    repeated IDAndTenant tenants = 1;
+}
+
+message IDAndTenant {
+    int64 id = 1;
+    Tenant tenant = 2;
+}
+
+message GetControlProxyRequest {
+  int64 id = 1;
+}
+
+message GetControlProxyResponse {
+  int64 id = 1;
+  string controlProxy = 2;
+}
+
+message CreateOrUpdateControlProxyRequest {
+  int64 id = 1;
+  string controlProxy = 2;
+}
+
+message CreateOrUpdateControlProxyResponse {
+}
+
+service TenantsService {
+    rpc GetAllTenants (Void) returns (TenantList) {}
+    rpc GetTenant (GetTenantRequest) returns (Tenant) {}
+    rpc CreateTenant (IDAndTenant) returns (Void) {}
+    rpc SetTenant (IDAndTenant) returns (Void) {}
+    rpc DeleteTenant (GetTenantRequest) returns (Void) {}
+    rpc GetControlProxy (GetControlProxyRequest) returns (GetControlProxyResponse) {}
+    rpc CreateOrUpdateControlProxy (CreateOrUpdateControlProxyRequest) returns (CreateOrUpdateControlProxyResponse) {}
+}
